SKIW-2 add ski area weather service

+++ src/main/kotlin/skiweather/model/weather/HourData.kt
@@ -7,5 +7,5 @@ data class HourData(
     val humidity: Int,
     val cloud: Int,
     val visKm: Double,
-    val uv: Int
+    val uv: Double
 )
